move sdktype lib
